requirements.txt
llama-index==0.10.0
llama-index-core==0.10.0
llama-index-embeddings-openai==0.1.0
llama-index-llms-openai==0.1.0
openai>=1.0.0
python-dotenv==1.0.0
rank-bm25>=0.2.2
